requirements.txt
apache-airflow
python-dotenv
requests
pandas
SQLAlchemy>=1.4.0